Commit:
Description: Overhauled code quality by adding CI pipeline, automated testing, dependency management, configuration, etc.

Changes:
- .github
  - Github Actions CI Pipeline
- .test_configs
  - Config files for pylint and flake8
- tests
  - Edited tests to check if a file exists before removing
- poetry.lock
  - Poetry managed file
- pyproject.toml
  - Modern package storage
- tox.ini
  - Automated testing

+++ tests/unit_tests/test_logger.py
@@ -7,11 +7,13 @@ import unittest
 import os
 from ...core.logger import Logger
 
+
 class TestLogger(unittest.TestCase):
     """
     Explanation:
         A testing suite for the features of logger.py
     """
+
     def setUp(self) -> None:
         """
         Explanation:
@@ -25,7 +27,8 @@ class TestLogger(unittest.TestCase):
         Explanation:
             Remove the test file
         """
-        os.remove(self.file_name)
+        if os.path.exists("./test.txt"):
+            os.remove(self.file_name)
 
     def test_singleton(self) -> None:
         """
